Changed some of the request fields to mirror the changes in the backend. Removed case_id and added product_id and purchase_order_id.
Added functionality, so now, when the user saves the updated request lines, those changes are actually made in the database.

Now, the picklist generator is being worked on. The attempt is to make a function that grabs boxes from the appropriate location for use in creation of a case. This means that boxes cannot be reused, which is the first big hurdle.