Continued to work on the purchase order editing function. It's a bit of a beast, so for now, the focus has been shifted to the request to process page.
New filters have been added to pull requested recipes, not cases. Additionally, cells automatically update when filled out, instead of the user having to select save on the bottom of the datatable.

Next, will work to ensure full functionality on the requests to process page, such as giving the user the ability to add new requests directly on the page.